requirements.txt
google-api-python-client==2.172.0
google-auth==2.40.3
ollama==0.5.1
streamlit==1.46.0
